fix code placing
