+++ src/extension.ts
@@ -49,7 +49,8 @@ const commands = [
     {name: "disableRepeatingDigits", func: updateEnableRepeatingDigits, value: false},
     {name: "enableRepeatingDigitsForUser", func: updateEnableRepeatingDigitsForUser, value: true},
     {name: "disableRepeatingDigitsForUser", func: updateEnableRepeatingDigitsForUser, value: false},
-    {name: "updateColorAtRepeatingDigitsForUser", func: getColorCodeAtRepeatingDigits, value: undefined},
+    {name: "updateColorAtRepeatingDigits", func: getColorCodeAtRepeatingDigits, value: undefined},
+    {name: "updateColorAtRepeatingDigitsForUser", func: getColorCodeAtRepeatingDigitsForUser, value: undefined},
     {name: "updateColorAtCenterOfRainbow", func: getColorCodeAtCenterOfRainbow, value: undefined},
     {name: "updateColorAtCenterOfRainbowForUser", func: getColorCodeAtCenterOfRainbowForUser, value: undefined},
     {name: "updateColorAtInactiveRowNumberForUser", func: getColorCodeAtInactiveRowNumberForUser, value: undefined},
